no casting

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -14,7 +14,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *ht;
-	int i;
+	unsigned long int i;
 
 	ht = malloc(sizeof(hash_node_t));
 	if (ht == NULL)
@@ -31,7 +31,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 		return (NULL);
 	}
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		ht->array[i] = NULL;
 	}
